Corrige tipos a long long para evitar overflow en contadores de nodos en BT, fusion main.cpp verificada funcionando

+++ src/BTSinPoda.hpp
@@ -8,13 +8,13 @@ class BacktrackerSinPoda {
 public:
     BacktrackerSinPoda(int n, int minLower, int minUpper, int minDigit, int minSymbol);
     void ejecutar();
-    int getContrasenasGeneradas() const { return contraseñasGeneradas; }
-    int getNodosGenerados() const { return nodosGenerados; }
+    long long getContrasenasGeneradas() const { return contraseñasGeneradas; }
+    long long getNodosGenerados() const { return nodosGenerados; }
     double getTiempoMs() const { return tiempoMs; }
     
 private:
     int usadoLower = 0, usadoUpper = 0, usadoDigit = 0, usadoSymbol = 0;
-    int contraseñasGeneradas = 0, nodosGenerados = 0, nodosVisitados = 0, nodosPodados = 0;
+    long long contraseñasGeneradas = 0, nodosGenerados = 0, nodosVisitados = 0, nodosPodados = 0;
     double tiempoMs = 0;
     const string alfabeto = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%";
     int n, minLower, minUpper, minDigit, minSymbol;
