+++ src/backtracking.hpp
@@ -8,15 +8,15 @@ class BacktrackerConPoda {
 public:
     BacktrackerConPoda(int n, int minLower, int minUpper, int minDigit, int minSymbol);
     void ejecutar();
-    int getContrasenasGeneradas() const { return contraseñasGeneradas; }
-    int getNodosGenerados() const { return nodosGenerados; }
-    int getNodosVisitados() const { return nodosVisitados; }
-    int getNodosPodados() const { return nodosPodados; }
+    long long getContrasenasGeneradas() const { return contraseñasGeneradas; }
+    long long getNodosGenerados() const { return nodosGenerados; }
+    long long getNodosVisitados() const { return nodosVisitados; }
+    long long getNodosPodados() const { return nodosPodados; }
     double getTiempoMs() const { return tiempoMs; }
     
 private:
     int usadoLower = 0, usadoUpper = 0, usadoDigit = 0, usadoSymbol = 0;
-    int contraseñasGeneradas = 0, nodosGenerados = 0, nodosVisitados = 0, nodosPodados = 0;
+    long long contraseñasGeneradas = 0, nodosGenerados = 0, nodosVisitados = 0, nodosPodados = 0;
     double tiempoMs = 0;
     const string alfabeto = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%";
     int n, minLower, minUpper, minDigit, minSymbol;
